Trait std_collection_traits::associated::AssociatedMap

pub trait AssociatedMap: AssociatedCollectionTrait + AssociatedContains {
    fn insert(
        &mut self, 
        key: Self::KeyType, 
        value: Self::ValueType
    ) -> Option<Self::ValueType>
    where
        Self: Owned;

    fn remove(&mut self, key: &Self::KeyType) -> Option<Self::ValueType>
    where
        Self: Owned;

    fn remove_entry(
        &mut self, 
        key: &Self::KeyType
    ) -> Option<(Self::KeyType, Self::ValueType)>
    where
        Self: Owned;

    fn get(&self, key: &Self::KeyType) -> Option<&Self::ValueType>;

    fn get_mut(&mut self, key: &Self::KeyType) -> Option<&mut Self::ValueType>
    where
        Self: Mutable;

    fn get_key_value(
        &self, 
        key: &Self::KeyType
    ) -> Option<(&Self::KeyType, &Self::ValueType)>;
}

Associated collections as key-value maps

Required methods

fn insert(
&mut self,
key: Self::KeyType,
value: Self::ValueType
) -> Option<Self::ValueType> where
Self: Owned,

Inserts a key-value pair into self.

fn remove(&mut self, key: &Self::KeyType) -> Option<Self::ValueType> where
Self: Owned,

Removes a key from self, returning the value at the key if the key was previously in the map.

fn remove_entry(
&mut self,
key: &Self::KeyType
) -> Option<(Self::KeyType, Self::ValueType)> where
Self: Owned,

Removes a key from self, returning the value at the key if the key was previously in the map.

fn get(&self, key: &Self::KeyType) -> Option<&Self::ValueType>

Returns a reference to the value corresponding to the key.

fn get_mut(&mut self, key: &Self::KeyType) -> Option<&mut Self::ValueType> where
Self: Mutable,

Returns a mutable reference to the value corresponding to the key.

fn get_key_value(
&self,
key: &Self::KeyType
) -> Option<(&Self::KeyType, &Self::ValueType)>

Returns the key-value pair corresponding to the supplied key.