Trait std_collection_traits::associated::AssociatedMapIter

pub trait AssociatedMapIter<S>: AssociatedMap {
    type TryInsertError;
    type KeyIter: Iterator<Item = &'a Self::KeyType>;
    type ValueIter: Iterator<Item = &'a Self::ValueType>;
    type ValueIterMut: Iterator<Item = &'a mut Self::ValueType>;
    fn try_insert(
        &mut self, 
        key: Self::KeyType, 
        value: Self::ValueType
    ) -> Result<&mut Self::ValueType, Self::TryInsertError>
    where
        Self: Owned;

    fn keys(&self) -> Self::KeyIter;

    fn values(&self) -> Self::ValueIter;

    fn values_mut(&mut self) -> Self::ValueIterMut
    where
        Self: Mutable;
}

Associated k-v maps operations

FIXME: Here we manually add S to handle lifetime requirements of HashMap<K, V, S>

Associated Types

type TryInsertError

Try insert error type.

Ignore std::error::Error because that only applies when K: Debug, V: Debug

type KeyIter: Iterator<Item = &'a Self::KeyType>

type ValueIter: Iterator<Item = &'a Self::ValueType>

type ValueIterMut: Iterator<Item = &'a mut Self::ValueType>

Required methods

fn try_insert(
&mut self,
key: Self::KeyType,
value: Self::ValueType
) -> Result<&mut Self::ValueType, Self::TryInsertError> where
Self: Owned,

Tries to insert a key-value pair into self, and returns a mutable reference to the value in the entry.

fn keys(&self) -> Self::KeyIter

Gets an iterator over the keys of self.

fn values(&self) -> Self::ValueIter

Gets an iterator over the values of self.

fn values_mut(&mut self) -> Self::ValueIterMut where
Self: Mutable,

Gets an mutable iterator over the values of self.