Trait std_collection_traits::associated::AssociatedMapOrd

pub trait AssociatedMapOrd: AssociatedMap {
    fn append(&mut self, other: &mut Self)
    where
        Self: Owned;

    fn split_off(&mut self, key: &Self::KeyType) -> Self
    where
        Self: Owned;

    fn first_key_value(&self) -> Option<(&Self::KeyType, &Self::ValueType)>;

    fn last_key_value(&self) -> Option<(&Self::KeyType, &Self::ValueType)>;

    fn pop_first(&mut self) -> Option<(Self::KeyType, Self::ValueType)>
    where
        Self: Owned;

    fn pop_last(&mut self) -> Option<(Self::KeyType, Self::ValueType)>
    where
        Self: Owned;
}

Associated collections that are ordered.

Required methods

fn append(&mut self, other: &mut Self) where
Self: Owned,

Moves all elements from other into self, leaving other empty.

fn split_off(&mut self, key: &Self::KeyType) -> Self where
Self: Owned,

Splits the collection into two at the given key.

fn first_key_value(&self) -> Option<(&Self::KeyType, &Self::ValueType)>

Returns the first key-value pair in the map.

fn last_key_value(&self) -> Option<(&Self::KeyType, &Self::ValueType)>

Returns the last key-value pair in the map.

fn pop_first(&mut self) -> Option<(Self::KeyType, Self::ValueType)> where
Self: Owned,

Removes the first value from self and returns it, if any.

fn pop_last(&mut self) -> Option<(Self::KeyType, Self::ValueType)> where
Self: Owned,

Removes the last value from self and returns it, if any.