Trait std_collection_traits::associated::AssociatedSet

pub trait AssociatedSet: CollectionTrait + Contains + Iterable {
    fn is_disjoint(&self, other: &Self) -> bool;

    fn is_subset(&self, other: &Self) -> bool;

    fn is_superset(&self, other: &Self) -> bool;

    fn take(&mut self, value: &Self::ElemType) -> Option<Self::ElemType>
    where
        Self: Owned;

    fn insert(&mut self, value: Self::ElemType) -> bool
    where
        Self: Owned;

    fn get(&self, value: &Self::ElemType) -> Option<&Self::ElemType>;

    fn remove(&mut self, value: &Self::ElemType) -> bool
    where
        Self: Owned;

    fn replace(&mut self, value: Self::ElemType) -> Option<Self::ElemType>
    where
        Self: Owned;
}

Associated collections as sets.

NOTE: Ignore all T: Borrow<Q> since we do not have way to parameterize bounds.

Required methods

fn is_disjoint(&self, other: &Self) -> bool

Returns true if self has no elements in common with other

fn is_subset(&self, other: &Self) -> bool

Returns true if self is a subset of another.

fn is_superset(&self, other: &Self) -> bool

Returns true if self is a superset of another.

fn take(&mut self, value: &Self::ElemType) -> Option<Self::ElemType> where
Self: Owned,

Removes and returns the value in self, if any, that is equal to the given one.

fn insert(&mut self, value: Self::ElemType) -> bool where
Self: Owned,

Adds a value to the set.

fn get(&self, value: &Self::ElemType) -> Option<&Self::ElemType>

Returns a reference to the value in the set, if any, that is equal to the given value.

fn remove(&mut self, value: &Self::ElemType) -> bool where
Self: Owned,

Removes a value from self. Returns whether the value was present in self.

fn replace(&mut self, value: Self::ElemType) -> Option<Self::ElemType> where
Self: Owned,

Adds a value to self, replacing the existing value, if any, that is equal to the given one. Returns the replaced value.