Trait std_collection_traits::associated::AssociatedSetOrd

pub trait AssociatedSetOrd: AssociatedSet {
    fn append(&mut self, other: &mut Self)
    where
        Self: Owned;

    fn first(&self) -> Option<&Self::ElemType>;

    fn pop_first(&mut self) -> Option<Self::ElemType>
    where
        Self: Owned;

    fn last(&self) -> Option<&Self::ElemType>;

    fn pop_last(&mut self) -> Option<Self::ElemType>
    where
        Self: Owned;

    fn split_off(&mut self, key: &Self::ElemType) -> Self
    where
        Self: Owned;
}

Associated collections as ordered sets

Required methods

fn append(&mut self, other: &mut Self) where
Self: Owned,

Moves all elements from other into self, leaving other empty.

fn first(&self) -> Option<&Self::ElemType>

Returns a reference to the first value in self, if any.

fn pop_first(&mut self) -> Option<Self::ElemType> where
Self: Owned,

Removes the first value from self and returns it, if any.

fn last(&self) -> Option<&Self::ElemType>

Returns a reference to the last value in self, if any.

fn pop_last(&mut self) -> Option<Self::ElemType> where
Self: Owned,

Removes the last value from self and returns it, if any.

fn split_off(&mut self, key: &Self::ElemType) -> Self where
Self: Owned,

Splits the collection into two at the given key.