Trait std_collection_traits::capacity::CapacityReserve

pub trait CapacityReserve: CapacityAware {
    type TryReserveError: Error;
    fn reserve(&mut self, additional: Self::CapacityType);

    fn try_reserve(
        &mut self, 
        additional: Self::CapacityType
    ) -> Result<(), Self::TryReserveError>;
}

Collections that can reserve capacity.

Associated Types

type TryReserveError: Error

Try reserve errors

Required methods

fn reserve(&mut self, additional: Self::CapacityType)

Reserves capacity for at least additional more elements to be inserted in the given Self.

fn try_reserve(
&mut self,
additional: Self::CapacityType
) -> Result<(), Self::TryReserveError>

Tries to reserve capacity for at least additional more elements to be inserted in the given Self.