Trait std_collection_traits::capacity::CapacityShrink

pub trait CapacityShrink: CapacityAware {
    fn shrink_to(&mut self, min_capacity: Self::CapacityType);

    fn shrink_to_fit(&mut self) { ... }
}

Collections that can shrink based on actual elements.

Required methods

fn shrink_to(&mut self, min_capacity: Self::CapacityType)

Shrinks the capacity of Self with a lower bound.

Provided methods

fn shrink_to_fit(&mut self)

Shrinks the capacity of Self as much as possible.