Module std_collection_traits::iter

Traits

AssociatedDrainFilter

Associated collections that can drain elements given a filter on k-v reference

AssociatedDrainFilterSet

Collections that can drain elements given a filter on immutable reference

AssociatedDrainFull

Collections that can drain all elements

AssociatedIterable

Collections that can iterate as sequence of (&K, &V).

AssociatedIterableMut

Collections that can iterate as sequence of (&K, &mut V).

AssociatedRange

Collections that can return a immutable range of elements based on key values

AssociatedRangeMut

Collections that can return a mutable range of elements based on key values

DrainFilter

Collections that can drain elements given a filter on mutable reference

DrainFull

Collections that can drain all elements

DrainRange

Collections that can drain a range of elements

Iterable

Collections that can iterate as sequence of &T.

IterableMut

Collections that can iterate as sequence of &mut T.

Range

Collections that can return a immutable range of elements

RangeMut

Collections that can return a mutable range of elements