Trait std_collection_traits::iter::AssociatedDrainFilterSet

pub trait AssociatedDrainFilterSet: CollectionTrait + Mutable {
    type DrainFilterIter: Iterator<Item = Self::ElemType>;
    fn drain_filter<'a, F: FnMut(&Self::ElemType) -> bool + 'a>(
        &'a mut self, 
        filter: F
    ) -> Self::DrainFilterIter;
}

Collections that can drain elements given a filter on immutable reference

Associated Types

type DrainFilterIter: Iterator<Item = Self::ElemType>

Required methods

fn drain_filter<'a, F: FnMut(&Self::ElemType) -> bool + 'a>(
&'a mut self,
filter: F
) -> Self::DrainFilterIter

Creates a draining iterator that removes the specified range in self and yields the removed items.

FIXME: Separate FnMut(&T) -> bool because BTreeSet::drain requires it. Otherwise, it can be FnMut(&mut T). Meanwhile, there is no easy way to convert FnMut(&T) -> bool to Fn(&K, &mut ()) -> bool and specify it at the associated type part/