Trait std_collection_traits::iter::AssociatedRangeMut

pub trait AssociatedRangeMut: AssociatedCollectionTrait + Mutable {
    type RangeIterMut: Iterator<Item = (&'a Self::KeyType, &'a mut Self::ValueType)>;
    fn range_mut<R: RangeBounds<Self::KeyType>>(
        &mut self, 
        range: R
    ) -> Self::RangeIterMut;
}

Collections that can return a mutable range of elements based on key values

Associated Types

type RangeIterMut: Iterator<Item = (&'a Self::KeyType, &'a mut Self::ValueType)>

Required methods

fn range_mut<R: RangeBounds<Self::KeyType>>(
&mut self,
range: R
) -> Self::RangeIterMut

Creates an iterator that covers the specified range in the self

NOTE: BTreeMap::range are more general.