Trait std_collection_traits::iter::Range

pub trait Range: CollectionTrait + ExactSized {
    type RangeIter: Iterator<Item = &'a Self::ElemType>;
    fn range<R: RangeBounds<Self::SizeType>>(&self, range: R) -> Self::RangeIter;
}

Collections that can return a immutable range of elements

Associated Types

type RangeIter: Iterator<Item = &'a Self::ElemType>

Required methods

fn range<R: RangeBounds<Self::SizeType>>(&self, range: R) -> Self::RangeIter

Creates an iterator that covers the specified range in the self