Trait std_collection_traits::sequential::RandomAccess

pub trait RandomAccess: Sequential + ExactSized {
    fn partition_point<P: FnMut(&Self::ElemType) -> bool>(
        &self, 
        pred: P
    ) -> Self::SizeType;

    fn get(&self, index: Self::SizeType) -> Option<&Self::ElemType>;

    fn rotate_left(&mut self, mid: Self::SizeType)
    where
        Self: Mutable,
    { ... }

    fn rotate_right(&mut self, mid: Self::SizeType)
    where
        Self: Mutable,
    { ... }

    fn swap(&mut self, i: Self::SizeType, j: Self::SizeType)
    where
        Self: Mutable,
    { ... }

    fn get_mut(&mut self, index: Self::SizeType) -> Option<&mut Self::ElemType>
    where
        Self: Mutable,
    { ... }

    fn insert(&mut self, index: Self::SizeType, value: Self::ElemType)
    where
        Self: Owned,
    { ... }

    fn swap_remove_back(
        &mut self, 
        index: Self::SizeType
    ) -> Option<Self::ElemType>
    where
        Self: Owned,
    { ... }
}

Collections that are randomly accessible in about O(1) time.

NOTE: complete definition requires TrustRandomAccess.

Required methods

fn partition_point<P: FnMut(&Self::ElemType) -> bool>(
&self,
pred: P
) -> Self::SizeType

Returns the index of the partition point according to the given predicate (the index of the first element of the second partition).

fn get(&self, index: Self::SizeType) -> Option<&Self::ElemType>

Provides a reference to the element at the given index.

Provided methods

fn rotate_left(&mut self, mid: Self::SizeType) where
Self: Mutable,

Rotates the double-ended queue mid places to the left.

fn rotate_right(&mut self, mid: Self::SizeType) where
Self: Mutable,

Rotates the double-ended queue mid places to the right.

fn swap(&mut self, i: Self::SizeType, j: Self::SizeType) where
Self: Mutable,

Swaps elements at indices i and j.

fn get_mut(&mut self, index: Self::SizeType) -> Option<&mut Self::ElemType> where
Self: Mutable,

Provides a mutable reference to the element at the given index.

fn insert(&mut self, index: Self::SizeType, value: Self::ElemType) where
Self: Owned,

Inserts an element at index within self, shifting all elements with indices greater than or equal to index towards the back.

fn swap_remove_back(&mut self, index: Self::SizeType) -> Option<Self::ElemType> where
Self: Owned,

Removes an element from anywhere in self and returns it, replacing it with the last element.