Trait std_collection_traits::sequential::Sequential

pub trait Sequential: CollectionTrait + ExactSized {
    fn back(&self) -> Option<&Self::ElemType>;

    fn back_mut(&mut self) -> Option<&mut Self::ElemType>
    where
        Self: Mutable,
    { ... }

    fn append(&mut self, other: &mut Self)
    where
        Self: Owned,
    { ... }

    fn push_back(&mut self, elt: Self::ElemType)
    where
        Self: Owned,
    { ... }

    fn pop_back(&mut self) -> Option<Self::ElemType>
    where
        Self: Owned,
    { ... }

    fn remove(&mut self, at: Self::SizeType) -> Option<Self::ElemType>
    where
        Self: Owned,
    { ... }

    fn split_off(&mut self, at: Self::SizeType) -> Self
    where
        Self: Owned + Sized,
    { ... }
}

Collections that behave like a sequence that can access from front to back.

Required methods

fn back(&self) -> Option<&Self::ElemType>

Provides a reference to the back element, or None if self is empty.

Provided methods

fn back_mut(&mut self) -> Option<&mut Self::ElemType> where
Self: Mutable,

Provides a mutable reference to the back element, or None if self is empty.

fn append(&mut self, other: &mut Self) where
Self: Owned,

Moves all elements from other into self, leaving other empty.

fn push_back(&mut self, elt: Self::ElemType) where
Self: Owned,

Appends an element to the back of self

fn pop_back(&mut self) -> Option<Self::ElemType> where
Self: Owned,

Removes the last element from self and returns it, or None if it is empty.

fn remove(&mut self, at: Self::SizeType) -> Option<Self::ElemType> where
Self: Owned,

Removes the element at the given index and returns it.

fn split_off(&mut self, at: Self::SizeType) -> Self where
Self: Owned + Sized,

Splits the collection into two at the given index.